fix: fix dashboard dto

+++ src/main/java/com/arctic/backend_for_arctic_team/metrics/model/dto/response/data_for_visualisation/SessionMetricsDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 @Builder
 public class SessionMetricsDto {
-    private Long session;
+    private Integer session;
     private String date;
     private String timeOfDay;
